refactor chargepoint
fix rebase

move files

fix
